+++ test/unit/clarity_ext/utility/test_hamilton_driver_file_reader.py
@@ -4,12 +4,11 @@
 import unittest
 import os
 import inspect
-from ...utility.dilute_filer_reader.hamilton_driver_file_reader import HamiltonReader
-from ...utility.dilute_filer_reader.hamilton_driver_file_reader import HamiltonColumnReference
+from clarity_ext.utility.hamilton_driver_file_reader import HamiltonReader, HamiltonColumnReference
 
 # TODO: Move the resource file closer to the corresponding test
 DRIVER_FILE_RELATIVE_PATH = os.path.join(os.path.dirname(__file__),
-                                         "../../mock/dilute_files",
+                                         "resources",
                                          "SX614_SX686_160308_4_J_Ham.txt")
